HypreErrorHandling: move out of gpuistl
put in new folder/namespace linalg/hypreinterface
